Added a simple layout for each Greyhound page
The layout is almost completely blank save for a footer with a
twitter button to share the current page.  Which allows each
Greyhound's page to be a little more unique.

Add a markdown file like the flowers.md example file to create a
new Greyhound page.
The layout variable needs to be greyhound, like in flowers.md
The title should be the name of the Greyhound featured
And the permalink should be a unique url path starting with /greyhounds/
Something close to the name of the would work best.