Added the events page.
This page shows a list of upcoming events for GPA, Central Texas.

I'd like to look into using Jekyll's new collections api to make
events work the same as posts.  That is creating a new event only
requires creating a new file in a _events directory and the events
page populating that file's contents within the list.