Update contact.md
Location change for monthly BOD meetings.  I'm not sure if I put everything in correctly or not.  Zach, can you double check it.  I know the address is right, and the new goo.gl shortened URL should point to the right spot.  I just want to make sure that it's going to show up online okay.