Move most greyhound info to greyhounds data file
We can now access data by key values in Jekyll, which means we can put almost all of the info
on a greyhound in the greyhound data file and access it in any markdown file.  This reduces the
amount of duplicate information, and can allow us to show more info on the Available page,
like cat safeness and/or sex.