Calculate a Greyhound's age label based on date of birth
Each greyhound in _data/greyhounds.yml can be given a date of birth field (dob) with a date formatted like so: YYYY-MM-DD
which will be used to create the age label for their profile page.
The greyhound's age will then automatically be updated when the site is rebuilt.

For available greyhounds, the age label will be how many years or months old a greyhound is.
e.g. "5 YEARS" or "8 MONTHS"

For adopted greyhounds, the age label will be the greyhound's birthday.
e.g. "Born 01 Jan 2000"

For deceased greyhounds, a date of death field (dod) will also be used to set the age label to a duration of life, or the greyhound's deathday
e.g. "01 Jan 2000 - 01 Jan 2010" or "Died 01 Jan 2010"