Change navbar to show dropdown menus on hover
This fixes a bug in IE where the carets would be off on desktops.  By
removing the carets in favor of a hover dropdown on desktop browsers the
item all line up correctly.

Also, this reverts the centering of menu items on mobile browsers,
because that just never looked quite right.